# frontend/src/lib/api/openapi.json
{"openapi":"3.1.0","info":{"title":"UniFi Device Control API","description":"HTTP API that mirrors the device management capabilities of the Streamlit dashboard.","version":"1.0.0"},"paths":{"/api/dashboard/summary":{"get":{"tags":["devices"],"summary":"Get Dashboard Summary","operationId":"get_dashboard_summary_api_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardSummary"}}}}}}},"/api/devices":{"get":{"tags":["devices"],"summary":"List Devices","operationId":"list_devices_api_devices_get","parameters":[{"name":"owner","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"title":"Owner"}},{"name":"locked","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Locked"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/owners":{"get":{"tags":["devices"],"summary":"List Owner Summaries","operationId":"list_owner_summaries_api_owners_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnersResponse"}}}}}}},"/api/owners/{owner_key}/verify-pin":{"post":{"tags":["devices"],"summary":"Verify Owner","operationId":"verify_owner_api_owners__owner_key__verify_pin_post","parameters":[{"name":"owner_key","in":"path","required":true,"schema":{"type":"string","title":"Owner Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPinRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPinResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/owners/{owner_key}/devices":{"get":{"tags":["devices"],"summary":"List Owner Devices","operationId":"list_owner_devices_api_owners__owner_key__devices_get","parameters":[{"name":"owner_key","in":"path","required":true,"schema":{"type":"string","title":"Owner Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/devices/lock":{"post":{"tags":["devices"],"summary":"Lock Devices","operationId":"lock_devices_api_devices_lock_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceActionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceActionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/owners/{owner_key}/lock":{"post":{"tags":["devices"],"summary":"Lock Owner Devices","operationId":"lock_owner_devices_api_owners__owner_key__lock_post","parameters":[{"name":"owner_key","in":"path","required":true,"schema":{"type":"string","title":"Owner Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerLockRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerLockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/clients/unregistered":{"get":{"tags":["devices"],"summary":"List Unregistered Clients","operationId":"list_unregistered_clients_api_clients_unregistered_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnregisteredClientsResponse"}}}}}}},"/api/clients/unregistered/lock":{"post":{"tags":["devices"],"summary":"Lock Unregistered Client","operationId":"lock_unregistered_client_api_clients_unregistered_lock_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleClientLockRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceActionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","description":"Simple readiness probe.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"DashboardSummary":{"properties":{"total_devices":{"type":"integer","minimum":0.0,"title":"Total Devices"},"locked_devices":{"type":"integer","minimum":0.0,"title":"Locked Devices"},"unlocked_devices":{"type":"integer","minimum":0.0,"title":"Unlocked Devices"},"owner_count":{"type":"integer","minimum":0.0,"title":"Owner Count"},"unknown_vendors":{"type":"integer","minimum":0.0,"title":"Unknown Vendors"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["total_devices","locked_devices","unlocked_devices","owner_count","unknown_vendors","generated_at"],"title":"DashboardSummary"},"DeviceActionRequest":{"properties":{"targets":{"items":{"$ref":"#/components/schemas/DeviceTarget"},"type":"array","minItems":1,"title":"Targets"},"unlock":{"type":"boolean","title":"Unlock","default":false}},"type":"object","required":["targets"],"title":"DeviceActionRequest"},"DeviceActionResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/DeviceActionResult"},"type":"array","title":"Results"}},"type":"object","required":["results"],"title":"DeviceActionResponse"},"DeviceActionResult":{"properties":{"mac":{"type":"string","title":"Mac"},"locked":{"type":"boolean","title":"Locked"},"status":{"type":"string","enum":["success","skipped","error"],"title":"Status"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["mac","locked","status"],"title":"DeviceActionResult"},"DeviceListResponse":{"properties":{"devices":{"items":{"$ref":"#/components/schemas/DeviceStatus"},"type":"array","title":"Devices"}},"type":"object","required":["devices"],"title":"DeviceListResponse"},"DeviceStatus":{"properties":{"name":{"type":"string","title":"Name"},"owner":{"type":"string","title":"Owner"},"type":{"type":"string","title":"Type"},"mac":{"type":"string","title":"Mac"},"locked":{"type":"boolean","title":"Locked"},"vendor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vendor"}},"type":"object","required":["name","owner","type","mac","locked"],"title":"DeviceStatus"},"DeviceTarget":{"properties":{"mac":{"type":"string","minLength":1,"title":"Mac"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"owner":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}},"type":"object","required":["mac"],"title":"DeviceTarget"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OwnerLockRequest":{"properties":{"unlock":{"type":"boolean","title":"Unlock","default":false}},"type":"object","title":"OwnerLockRequest"},"OwnerLockResponse":{"properties":{"owner":{"type":"string","title":"Owner"},"processed":{"type":"integer","title":"Processed"},"results":{"items":{"$ref":"#/components/schemas/DeviceActionResult"},"type":"array","title":"Results"}},"type":"object","required":["owner","processed","results"],"title":"OwnerLockResponse"},"OwnerSummary":{"properties":{"key":{"type":"string","title":"Key"},"display_name":{"type":"string","title":"Display Name"},"total_devices":{"type":"integer","title":"Total Devices"},"locked_devices":{"type":"integer","title":"Locked Devices"},"unlocked_devices":{"type":"integer","title":"Unlocked Devices"}},"type":"object","required":["key","display_name","total_devices","locked_devices","unlocked_devices"],"title":"OwnerSummary"},"OwnersResponse":{"properties":{"owners":{"items":{"$ref":"#/components/schemas/OwnerSummary"},"type":"array","title":"Owners"}},"type":"object","required":["owners"],"title":"OwnersResponse"},"SingleClientLockRequest":{"properties":{"mac":{"type":"string","title":"Mac"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"owner":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"unlock":{"type":"boolean","title":"Unlock","default":false}},"type":"object","required":["mac"],"title":"SingleClientLockRequest"},"UnregisteredClient":{"properties":{"name":{"type":"string","title":"Name"},"mac":{"type":"string","title":"Mac"},"ip":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip"},"vendor":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vendor"},"last_seen":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Seen"},"locked":{"type":"boolean","title":"Locked"}},"type":"object","required":["name","mac","locked"],"title":"UnregisteredClient"},"UnregisteredClientsResponse":{"properties":{"clients":{"items":{"$ref":"#/components/schemas/UnregisteredClient"},"type":"array","title":"Clients"}},"type":"object","required":["clients"],"title":"UnregisteredClientsResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyPinRequest":{"properties":{"pin":{"type":"string","maxLength":64,"minLength":1,"title":"Pin"}},"type":"object","required":["pin"],"title":"VerifyPinRequest"},"VerifyPinResponse":{"properties":{"valid":{"type":"boolean","title":"Valid"}},"type":"object","required":["valid"],"title":"VerifyPinResponse"}}}}